iot_control_platform/requirements.txt
Django
paho-mqtt
django-cors-headers
django-extensions
djangorestframework
djangorestframework-simplejwt
mysqlclient
gunicorn